pyproject.toml
[project]
name = "my-ai-news-aggregator"
version = "0.1.0"
description = "An AI-powered news agent that summarizes and categorizes news articles from various sources and sends daily digest to use."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "docling>=2.61.2",
    "feedparser>=6.0.12",
    "google-api-python-client>=2.190.0",
    "google-genai>=1.65.0",
    "markdown>=3.7.0",
    "markdownify>=0.11.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "youtube-transcript-api>=1.2.3",
]
